+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update to git2 0.15 which uses gits `safe.directory` logic.
+- Improve example-config
+
+### Removed
+
+- `check --rewrite-yaml` and `check --print-yaml`
 
 ## [0.19.0] - 2022-05-11